requirements.txt
Flask~=2.2.3
gunicorn==20.1.0
requests==2.28.2
PyJWT==2.6.0
cryptography==40.0.1
python-dateutil==2.8.2
